pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "This is a RAG pipeline."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb==0.5.18",
    "google-generativeai==0.8.3",
    "matplotlib==3.9.2",
    "paddleocr==2.7.3",
    "pdf2image==1.17.0",
    "pypdf2==3.0.1",
    "pypdfium2==4.30.1",
    "pytesseract==0.3.13",
    "python-dotenv==1.0.1",
    "ray==2.37.0",
    "scikit-learn==1.5.2",
    "streamlit==1.39.0",
    "networkx",
    "setuptools",
    "sentence-transformers",
    "python-multipart>=0.0.20",
]
